adjusting UserService, interface and UserDto
