+++ ZventsApi/DTOs/User/UserListDto.cs
@@ -1,7 +1,10 @@
-public class UserListDto
+namespace ZventsApi.DTOs.User
 {
-    public Guid Id { get; set; }
-    public string Name { get; set; }
-    public string UserName { get; set; }
-    public DateTime CreatedDate { get; set; }
+    public class UserListDto
+    {
+        public Guid Id { get; set; }
+        public string Name { get; set; }
+        public string UserName { get; set; }
+        public DateTime CreatedDate { get; set; }
+    }
 }
